SpamQuarantine: simplify download button by binding its href config
This avoids all the strange click event hack. We already have a
button where the users needs to click on, so just use that.

A viewModel with a downloadMailURL formula to assemble the href
allows us to just bind the "href" config from the Download button,
and be done with it..

refactors commit 7ad0de10d630f628b15969ef1478b9276e46197b

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>