Revert Adding of before_queue_filtering checkbox
This partially reverts commit 28c12d957fa1e4397154f66eb097c9639ced2b6d.

Given that the before queue filtering currently breaks pmg-log-tracker, and
the Tracking Center is one of the central features of PMG, we should
release the before queue support as experimental feature for now.

That way users who really want to use it are a bit more aware, that
log-tracking will currently not work on the GUI.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>