add PMG 5.2 End-of-Life notice
based on the one from PVE, but setting the width explicit is required
here.
since Mailgateway 5.2 goes out of support in July, add a notice at
the top with a link to the faq, where the EOL dates are and also the
upgrade process is described and linked in short.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>