buildsys: fix makefile dependencies
To ensure relevant stuff is always rebuild include the CSS and image
files in the install dependencies.

Further, mark the meta targets for building the pmgmanagerlib normal
and mobile versions as PHONY, as they only re-direct the make to the
respective makefile we need to ensure it always does, as else it
circumvents the actual dependency declaration of the responsible
targets.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>