quarantine: restore behavior of selecting next in list after action
This fixes a UX regression stemming from the recently introduced
commit 9aed379a700a0bc3bc26fa32d43a5bab75eb206b
"spam quarantine: avoid full store reload on single-element actions"

A load restored the saved selection after load complete, with the
neat and very useful feature to jump to the next in line if the old
one did not exist anymore.

Call this now factored out code part ourself in the fast-path.

This was reported in the forum:
https://forum.proxmox.com/threads/proxmox-mail-gateway-6-4-released.86760/#post-380971

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>