sys config: move network to bottom
Like we did in PBS, which has a very similar GUI structure as PMG.

Rationale is that the network panel is flexed and thus normally the
biggest, while DNS and time are fixed and short, and an inverted
pyramid, height-wise, looks a bit off as layout.

Also network will show the pending changes at the bottom, which is
nicer if its really at the bottom of the view.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>