quarantine: mail display: show receiver information
Showing the sender and subject is sufficient in the case of the
spam-quarantine - when viewed by a single user.

If an admin views any of the quarantines the information of the
recipient should be helpful in general, but is needed in case a mail
was sent to multiple recipients.

Reported in our community forum:
https://forum.proxmox.com/threads/.103276/

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>