spam quarantine controller: fix args when calling the parent multiSelect
arguments needs to be an array, often we call with all of ours, iow.
using `arguments` which itself is an array already.

Both, manual construction and just a passing `arguments` have their
advantage and disadvantage, e.g., if either the inheriting or
bequesting class becomes more specialized, but most of the time using
`arguments` is more future proof.

Fixes: e66c888 ("quarantine: refactor spamquarantine controller")
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>