mail filter: properly html-encode rule names
Rule names were not HTML-encoded in the rule overview. While some
encoding did happen, it was only on the first load for the info panel
on the right hand side, which was overriden as soon as the edit
window was opened/closed again.

The variable binding now gets HTML-encoded automatically, and the
renderer for rule names has been set to Ext.htmlEncode.

Signed-off-by: Leo Nunner <l.nunner@proxmox.com>
 [ T: rework commit message a bit (mostly subject) ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>